refactor: Move Navigation into ui, for ui-related components
- Forgot to adjust the actual component name for FetchVehicleData from App.
