+++ src/pages/fetch_vehicle_data/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import Navigation from '../../components/layout/Navigation';
+import Navigation from '../../components/ui/Navigation';
 import CarFetcher from '../../components/cars/CarFetcher';
 
-function App() {
+function FetchVehicleData() {
   return (
     <div className="bg-white min-h-screen">
       <Navigation />
@@ -12,4 +12,4 @@ function App() {
   );
 }
 
-export default App;
+export default FetchVehicleData;
